geo-api/requirements.txt
psycopg2-binary==2.8.4
python-dotenv==0.12.0
python-rapidjson==0.9.1
pytest==5.4.1
pytest-mock==3.0.0
Shapely==1.7.0
Pillow==7.1.1
requests==2.23.0
PyJWT==1.7.1
pandas==1.0.3
scikit-learn==0.23.0
tensorflow==2.2.0
sqlalchemy==1.3.17
singleton-decorator==1.0.0
bcrypt==3.1.7
flask==1.1.2
Flask-JWT-Extended==3.24.1
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.3
GeoAlchemy2==0.8.3
Flask-Migrate==2.5.3
sqlalchemy_mixins==1.2.1
flask-accept==0.0.6
shapely-geojson==0.0.1
python-magic==0.4.18; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'